feat(component-progress): Routes & firebase integration
