+++ package.json
@@ -3,12 +3,13 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@date-io/date-fns": "^1.3.13",
+    "@date-io/date-fns": "^2.10.8",
     "@material-ui/core": "^4.11.3",
     "@material-ui/icons": "^4.11.2",
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
+    "date-fns": "^2.0.0-beta.5",
     "firebase": "^8.2.9",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
@@ -46,6 +47,7 @@
   },
   "devDependencies": {
     "@material-ui/icons": "^4.11.2",
+    "@material-ui/pickers": "^3.2.10",
     "husky": "^5.0.9",
     "pinst": "^2.1.4"
   }
