+++ src/App.js
@@ -1,7 +1,7 @@
 import React from "react";
 import './App.css';
 import {AuthProvider} from "./services/Auth";
-import {LoginPage} from "./pages";
+import {LoginPage,RegisterProjectPage} from "./pages";
 import {BrowserRouter as Router, Route, Switch} from "react-router-dom"
 
 function App() {
@@ -11,6 +11,7 @@ function App() {
                 <AuthProvider>
                     <Switch>
                         <Route path="/login" component={LoginPage}/>
+                        <Route path="/register-project" component={RegisterProjectPage}/>
                     </Switch>
                 </AuthProvider>
             </Router>
