+++ src/components/Register/RegisterForm.js
@@ -1,15 +1,15 @@
 import React from 'react'
-import Register from './Register'
+import Register from '../../pages/RegisterPage/Register'
 
 
 
 const RegisterForm = () => {
     return (
-        
+
         <div>
             <Register/>
         </div>
-            
+
     )
 }
