+++ src/components/index.js
@@ -1,4 +1,3 @@
 export * from './ButtonCustom/Button'
 export * from './Form/Form';
 export * from './Progressbar/ProgressBar';
-export * from './ProjectForm/MultiStepForm';
