+++ src/pages/RegisterProjectPage/steps.js
