+++ src/pages/index.js
@@ -1,1 +1,2 @@
 export * from './LoginPage/index'
+export * from './RegisterProjectPage/index'
